backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pyjwt>=2.10.1",
    "whitenoise>=6.12.0",
]
